refactor(logger): extract shared @quackback/logger package
Move the logger implementation (pino, redaction, OpenTelemetry trace-context
mixin) and the AsyncLocalStorage request context into a shared
@quackback/logger workspace package so other packages can log within the same
request scope (one shared ALS instance).

- apps/web lib/server/{logger,log-context}.ts are now thin re-exports; the web
  wrapper fixes service_name to quackback-web, so the ~200 call sites and the
  unit tests are unchanged
- Migrate @quackback/email runtime send-path logs (transporter init, sent,
  failed) to the structured logger; recipient addresses are never logged. The
  dev-mode email preview boxes and the db CLI scripts stay on console as
  intentional human-facing output
- Extend the client logger stub and Vite alias to also cover @quackback/logger
  so pino and node:async_hooks never enter the browser bundle

Verified: typecheck, lint, full unit suite (4958 passing), production build
with a clean client bundle.
