+++ apps/web/package.json
@@ -33,6 +33,7 @@
     "@quackback/db": "workspace:*",
     "@quackback/email": "workspace:*",
     "@quackback/ids": "workspace:*",
+    "@quackback/logger": "workspace:*",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-avatar": "^1.1.11",
     "@radix-ui/react-checkbox": "^1.3.3",
