+++ apps/web/src/lib/server/log-context.ts
@@ -1,46 +1,8 @@
 /**
- * Per-request log context, carried via AsyncLocalStorage.
+ * Per-request log context — re-exported from the shared @quackback/logger
+ * package so the web app, @quackback/db and @quackback/email all share one
+ * AsyncLocalStorage instance (and therefore one request_id/tenant_id scope).
  *
- * A single store holds request-scoped identity so the logger can stamp every
- * line (request_id, tenant_id, ...) without passing a logger down the call
- * stack. Open the scope once in the global request middleware; enrich it with
- * setLogContext() once auth/tenant resolution runs.
- *
- * Server-only: imports node:async_hooks. Lives under lib/server/ and must never
- * be imported from client/isomorphic modules.
+ * Server-only: the underlying module imports node:async_hooks.
  */
-import { AsyncLocalStorage } from 'node:async_hooks'
-
-export interface LogContext {
-  /** Correlation id for the request; from x-request-id or generated. */
-  request_id: string
-  /** Low-cardinality route label, e.g. "GET /api/posts". */
-  route?: string
-  /** Workspace/tenant the request belongs to (high-cardinality; body field). */
-  tenant_id?: string
-  /** Authenticated user, once resolved. */
-  user_id?: string
-  /** Room for additional ambient fields without a type change. */
-  [key: string]: unknown
-}
-
-const storage = new AsyncLocalStorage<LogContext>()
-
-/** The active request context, or undefined outside a request scope. */
-export function getLogContext(): LogContext | undefined {
-  return storage.getStore()
-}
-
-/** Run `fn` with `context` as the ambient log context for its async subtree. */
-export function runWithLogContext<T>(context: LogContext, fn: () => T): T {
-  return storage.run(context, fn)
-}
-
-/**
- * Merge fields into the active context (e.g. tenant_id/user_id discovered after
- * the scope opened). No-op when called outside a request scope.
- */
-export function setLogContext(partial: Partial<LogContext>): void {
-  const store = storage.getStore()
-  if (store) Object.assign(store, partial)
-}
+export { getLogContext, runWithLogContext, setLogContext, type LogContext } from '@quackback/logger'
