+++ packages/email/package.json
@@ -12,6 +12,7 @@
     "./anon": "./src/anon.ts"
   },
   "dependencies": {
+    "@quackback/logger": "workspace:*",
     "@react-email/components": "^1.0.11",
     "nodemailer": "^8.0.4",
     "react": "^19.2.4",
